incoming v5 good but navigate not working

+++ app/src/main/java/com/example/smsapp/ui/incoming/v4/IncomingMessageTextV4.kt
@@ -28,7 +28,7 @@ fun IncomingMessageTextV4(
         Spacer(Modifier.height(8.dp))
 
         Text(
-            SmartTimeFormatter.format(sms.date),
+            SmartTimeFormatter.format(sms.date.toLong()),
             style = MaterialTheme.typography.labelSmall,
             color = MaterialTheme.colorScheme.onSurfaceVariant
         )
